crash fix

+++ app/src/main/java/com/shubham/bookreader/model/RetailPrice.kt
@@ -1,6 +1,6 @@
 package com.shubham.bookreader.model
 
 data class RetailPrice(
-    val amountInMicros: Int,
+    val amountInMicros: Long,
     val currencyCode: String
 )
